requirements.txt
pandas
spacy
google-generativeai
google-api-python-client